# api.yml
#Test with: http://editor.swagger.io/

swagger: '2.0'
info:
  version: 1.0.0
  title: Microservicio para Commentarios
  description: Microservicio para crear uno o más comentarios a una imagen del proyecto
  license:
    name: Krajono/Comments
host: 192.168.99.101:4100
basePath: /comments_ms/
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /comments:
    post:
      summary: Crea un nuevo comentario
      operationId: createComment
      parameters:
        - name: message
          in: body
          required: true
          description: Comment Object
          schema:
            $ref: '#/definitions/CommentInput'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    get:
      summary: Regresa todos los comentarios
      operationId: getComments
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/CommentList'
        '406':
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'

  /comments/{code}:
    get:
      summary: Regresa un comentario con el código dado
      operationId: getCommentByCode
      parameters:
        - name: code
          in: path
          description: El código del comentario a ser retornado
          required: true
          type: integer
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Comment'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
            description: Not Acceptable (Invalid Params)
            schema:
              $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'

    put:
      summary: Actualiza un comentario con el código dado
      operationId: updateComment
      parameters:
        - name: code
          in: path
          description: El código del comentario para ser actualizado
          required: true
          type: integer
        - name: message
          in: body
          required: true
          description: Comment Object
          schema:
            $ref: '#/definitions/CommentInput'
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
            description: Not Acceptable (Invalid Params)
            schema:
              $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'

    delete:
      summary: Eliminar un comentario con el código dado
      operationId: deleteComment
      parameters:
        - name: code
          in: path
          description: El código del comentario a ser eliminado
          required: true
          type: integer
      responses:
        '200':
          description: Ok
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        '406':
            description: Not Acceptable (Invalid Params)
            schema:
              $ref: '#/definitions/ErrorModel'
        '500':
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'

definitions:

  # Comments
  Comment:
    type: object
    required:
      - message
      - image_id
      - version_id
    properties:
      code:
        type: integer
      message:
        type: string
      image_id:
        type: integer
      version_id:
        type: integer
      likes:
        type: integer
      created_by:
        type: integer
      created_at:
        type: integer

  CommentInput:
    type: object
    required:
      - message
      - image_id
      - version_id
    properties:
      message:
        type: string
      image_id:
        type: integer
      version_id:
        type: integer
      likes:
        type: integer

  CommentList:
    type: object
    required:
      - total
      - list
    properties:
      total:
        type: number
      list:
        type: array
        items:
          $ref: '#/definitions/Comment'

# Errors
  ErrorModel:
    type: object
    required:
      - message
      - code
    properties:
      message:
        type: string
      code:
        type: number
      description:
        type: object